% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{get_nsduhus_available_years}
\alias{get_nsduhus_available_years}
\title{Print years available for download with the package.}
\usage{
get_nsduhus_available_years()
}
\value{
A character vector of years available.
}
\description{
Print years available for download with the package.
}
