% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{load_nsduhus}
\alias{load_nsduhus}
\title{Load downloaded U.S. NSDUH data files.}
\usage{
load_nsduhus(years = NULL, read_from_wd = FALSE)
}
\arguments{
\item{years}{(optional) A character vector of years to parse.}

\item{read_from_wd}{(optional) If TRUE, data files are loaded from
the working directory.}
}
\value{
A list of data frames.
}
\description{
Load downloaded U.S. NSDUH data files from \insertCite{nsduh}{nsduhus}.
}
\references{
\insertRef{nsduh}{nsduhus}
}
