% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{open_nsduhus_qshowcards}
\alias{open_nsduhus_qshowcards}
\title{Open the questionnaire showcards from source.}
\usage{
open_nsduhus_qshowcards(year)
}
\arguments{
\item{year}{A character string for a year (e.g., 2019).}
}
\description{
Open the questionnaire showcards from \insertCite{nsduh}{nsduhus}
for a specific year on a browser.
}
\references{
\insertRef{nsduh}{nsduhus}
}
